refactor: extract QuaiPayBottomSheetModal component
